fix(api): cast features to f32 before ONNX inference
The model expects tensor(float) but we were passing tensor(double),
causing a runtime error on every prediction request.

+++ api/src/services/rank_predictor.rs
@@ -192,7 +192,7 @@ fn run_session(
     session: &Mutex<Session>,
     features: &[f64; N_FEATURES],
 ) -> Result<f32, RankPredictorError> {
-    let input = Array2::from_shape_fn((1, N_FEATURES), |(_, j)| features[j]);
+    let input = Array2::from_shape_fn((1, N_FEATURES), |(_, j)| features[j] as f32);
     let tensor = Tensor::from_array(input)?;
     let mut guard = session
         .lock()
